Ajuste pruebas unitarias #15
