+++ test/assets/requests/BodyGen5.json
@@ -1,0 +1,8 @@
+{
+    "Nombre": "ssdfdfgsdfg",
+    "Activo": false,
+    "Formato": "{}",
+    "FechaCreacion": "2019-08-23T08:42:45.426405-05:00",
+    "FechaModificacion": "2019-08-23T08:42:45.426405-05:00",
+    "Descripcion": ""
+}
