+++ test/assets/requests/BodyRec4.json
@@ -1,0 +1,11 @@
+{
+    "Id": 100,
+    "Nombre": "asdfasd",
+    "CodigoAbreviacion": "",
+    "NumeroOrden": 0,
+    "Activo": false,
+    "Formato": "{}",
+    "FechaCreacion": "2019-08-23T08:42:45.426405-05:00",
+    "FechaModificacion": "2019-08-23T08:42:45.426405-05:00",
+    "Descripcion": ""
+}
