+++ test/assets/requests/Nt3.json
@@ -1,6 +1,6 @@
 {
     "Nombre": "",
-    "Descripcion": "",
+    "Descripcion": 0,
     "CodigoAbreviacion": 0,
     "Activo": false,
     "NumeroOrden": 0,
