+++ test/assets/requests/Nt5.json
@@ -1,5 +1,5 @@
 {
-    "Nombre": "",
+    "Nombre": 0,
     "Descripcion": "",
     "CodigoAbreviacion": "",
     "Activo": false,
