+++ test/movimientos_arka_crud.go
@@ -1,32 +1,26 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
-	"mime/multipart"
-	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"reflect"
 	"strconv"
 	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/DATA-DOG/godog"
-	"github.com/DATA-DOG/godog/colors"
+	"github.com/cucumber/godog"
+	"github.com/cucumber/godog/colors"
 	"github.com/astaxie/beego"
-	"github.com/xeipuuv/gojsonschema"
 
 	"github.com/udistrital/utils_oas/request"
 )
+
 
 //@opt opciones de godog
 var opt = godog.Options{Output: colors.Colored(os.Stdout)}
@@ -89,7 +83,7 @@ func run_bee() {
 	var resultado map[string]interface{}
 	
 
-	parametros := 	"MOVIMIENTOS_ARKA_CRUD_PGPORT=" + beego.AppConfig.String("PGport") +"MOVIMIENTOS_ARKA_CRUD_HTTP_PORT=" + beego.AppConfig.String("httpport") + " MOVIMIENTOS_ARKA_CRUD_PGUSER=" + beego.AppConfig.String("PGuser") + " MOVIMIENTOS_ARKA_CRUD_PGPASS=" + beego.AppConfig.String("PGpass") + " MOVIMIENTOS_ARKA_CRUD_PGURLS=" + beego.AppConfig.String("PGurls") + " MOVIMIENTOS_ARKA_CRUD_PGDB=" + beego.AppConfig.String("PGdb") + " MOVIMIENTOS_ARKA_CRUD_PGSCHEMA=" + beego.AppConfig.String("PGschemas") + " bee run"
+	parametros := "MOVIMIENTOS_ARKA_CRUD_PGPORT=" + beego.AppConfig.String("PGport") +" MOVIMIENTOS_ARKA_CRUD_HTTP_PORT=" + beego.AppConfig.String("httpport") + " MOVIMIENTOS_ARKA_CRUD_PGUSER=" + beego.AppConfig.String("PGuser") + " MOVIMIENTOS_ARKA_CRUD_PGPASS=" + beego.AppConfig.String("PGpass") + " MOVIMIENTOS_ARKA_CRUD_PGURLS=" + beego.AppConfig.String("PGurls") + " MOVIMIENTOS_ARKA_CRUD_PGDB=" + beego.AppConfig.String("PGdb") + " MOVIMIENTOS_ARKA_CRUD_PGSCHEMA=" + beego.AppConfig.String("PGschemas") + " bee run"
 	file, err := os.Create("script.sh")
 	if err != nil {
 		log.Fatal("Cannot create file", err)
@@ -137,7 +131,7 @@ func init() {
 
 //@TestMain para realizar la ejecucion con el comando go test ./test
 func TestMain(m *testing.M) {
-	
+	// init()
 	status := godog.RunWithOptions("godogs", func(s *godog.Suite) {
 		FeatureContext(s)
 	}, godog.Options{
@@ -176,220 +170,4 @@ func gen_files() {
 	ioutil.WriteFile("./assets/requests/BodyGen4.json", rankingsJson, 0644)
 }
 
-/*------------------------------
-  ---- Ejecución de pruebas ----
-  ------------------------------*/
 
-//@AreEqualJSON comparar dos JSON si son iguales retorna true de lo contrario false
-func AreEqualJSON(s1, s2 string) (bool, error) {
-
-	var o1 interface{}
-	var o2 interface{}
-
-	var err error
-	err = json.Unmarshal([]byte(s1), &o1)
-	if err != nil {
-		return false, fmt.Errorf("Error mashalling string 1 :: %s", err.Error())
-	}
-	err = json.Unmarshal([]byte(s2), &o2)
-	if err != nil {
-		return false, fmt.Errorf("Error mashalling string 2 :: %s", err.Error())
-	}
-
-	return reflect.DeepEqual(o1, o2), nil
-}
-
-//@toJson convierte string en JSON
-func toJson(p interface{}) string {
-
-	bytes, err := json.Marshal(p)
-	if err != nil {
-		fmt.Println(err.Error())
-		os.Exit(1)
-	}
-
-	return string(bytes)
-}
-
-//@getPages convierte en un tipo el json
-func getPages(ruta string) []byte {
-
-	raw, err := ioutil.ReadFile(ruta)
-	if err != nil {
-		fmt.Println(err.Error())
-		os.Exit(1)
-	}
-
-	var c []byte
-	c = raw
-	return c
-}
-
-//@iSendRequestToWhereBodyIsJson realiza la solicitud a la API
-func iSendRequestToWhereBodyIsJson(method, endpoint, bodyreq string) error {
-
-	var url string
-
-	if method == "GET" || method == "POST" {
-		url = "http://" + beego.AppConfig.String("PGurls") + ":" + beego.AppConfig.String("httpport") + endpoint
-
-	} else {
-		if method == "PUT" || method == "DELETE" {
-			str := strconv.FormatFloat(Id, 'f', 5, 64)
-			url = "http://" + beego.AppConfig.String("PGurls") + ":" + beego.AppConfig.String("httpport") + endpoint + "/" + str
-
-		}
-	}
-	if method == "GETID" {
-		method = "GET"
-		str := strconv.FormatFloat(Id, 'f', 0, 64)
-		url = "http://" + beego.AppConfig.String("PGurls") + ":" + beego.AppConfig.String("httpport") + endpoint + "/" + str
-
-	}
-	if method == "DELETE" {
-		str := strconv.FormatFloat(Id, 'f', 0, 64)
-		url = "http://" + beego.AppConfig.String("PGurls") + ":" + beego.AppConfig.String("httpport") + endpoint + "/" + str
-		resDelete = "{\"Id\":" + str + "}"
-		ioutil.WriteFile("./assets/responses/Ino.json", []byte(resDelete), 0644)
-
-	}
-
-	pages := getPages(bodyreq)
-
-	req, err := http.NewRequest(method, url, bytes.NewBuffer(pages))
-	req.Header.Set("Content-Type", "application/json")
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	bodyr, _ := ioutil.ReadAll(resp.Body)
-
-	resStatus = resp.Status
-	resBody = bodyr
-
-	if method == "POST" && resStatus == "201 Created" {
-		ioutil.WriteFile("./assets/requests/BodyRec2.json", resBody, 0644)
-		json.Unmarshal([]byte(bodyr), &savepostres)
-		Id = savepostres["Id"].(float64)
-
-	}
-	return nil
-
-}
-
-//@theResponseCodeShouldBe valida el codigo de respuesta
-func theResponseCodeShouldBe(arg1 string) error {
-	if resStatus != arg1 {
-		return fmt.Errorf("se esperaba el codigo de respuesta .. %s .. y se obtuvo el codigo de respuesta .. %s .. ", arg1, resStatus)
-	}
-	return nil
-}
-
-//@theResponseShouldMatchJson valida el JSON de respuesta
-func theResponseShouldMatchJson(arg1 string) error {
-	div := strings.Split(arg1, "")
-
-	pages := getPages(arg1)
-	//areEqual, _ := AreEqualJSON(string(pages), string(resBody))
-	if div[13] == "V" {
-		schemaLoader := gojsonschema.NewStringLoader(string(pages))
-		documentLoader := gojsonschema.NewStringLoader(string(resBody))
-		result, err := gojsonschema.Validate(schemaLoader, documentLoader)
-		if err != nil {
-			fmt.Println(err.Error())
-			os.Exit(1)
-		}
-		if result.Valid() {
-			return nil
-		} else {
-			return fmt.Errorf("Errores : %s", result.Errors())
-
-			return nil
-		}
-	}
-	if div[13] == "I" {
-		areEqual, _ := AreEqualJSON(string(pages), string(resBody))
-		if areEqual {
-			return nil
-		} else {
-			return fmt.Errorf(" se esperaba el body de respuesta %s y se obtuvo %s", string(pages), resBody)
-		}
-
-	}
-	return nil
-}
-
-// @iSendRequestToWhereBodyIsMultipartformdataWithThisParamsAndTheFileLocatedAt realiza la solicitud a la API
-func iSendRequestToWhereBodyIsMultipartformdataWithThisParamsAndTheFileLocatedAt(method, endpoint, bodyreq string, filename string, bodyfile string) error {
-
-	var url string
-
-	if method == "GET" || method == "POST" {
-		url = "http://" + beego.AppConfig.String("appurl") + ":" + beego.AppConfig.String("httpport") + endpoint
-	} else {
-		if method == "PUT" || method == "DELETE" {
-			url = "http://" + beego.AppConfig.String("appurl") + ":" + beego.AppConfig.String("httpport") + endpoint
-		}
-	}
-
-	extraParams := getPages(bodyreq)
-	var params map[string]string
-	err := json.Unmarshal(extraParams, &params)
-	if err != nil {
-		return err
-	}
-
-	path, _ := os.Getwd()
-	path += "/"
-	path += bodyfile
-
-	file, err := os.Open(path)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	body := &bytes.Buffer{}
-	writer := multipart.NewWriter(body)
-	part, err := writer.CreateFormFile(filename, filepath.Base(path))
-	if err != nil {
-		return err
-	}
-	_, err = io.Copy(part, file)
-
-	for key, val := range params {
-		_ = writer.WriteField(key, val)
-	}
-	err = writer.Close()
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequest(method, url, body)
-	req.Header.Set("Content-Type", writer.FormDataContentType())
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	bodyr, _ := ioutil.ReadAll(resp.Body)
-
-	resStatus = resp.Status
-	resBody = bodyr
-
-	return nil
-}
-
-func FeatureContext(s *godog.Suite) {
-	s.Step(`^I send "([^"]*)" request to "([^"]*)" where body is multipart\/form-data with this params "([^"]*)" and the file "([^"]*)" located at "([^"]*)"$`, iSendRequestToWhereBodyIsMultipartformdataWithThisParamsAndTheFileLocatedAt)
-	s.Step(`^I send "([^"]*)" request to "([^"]*) where body is json "([^"]*)"$`, iSendRequestToWhereBodyIsJson)
-	s.Step(`^the response code should be "([^"]*)"$`, theResponseCodeShouldBe)
-	s.Step(`^the response should match json "([^"]*)"$`, theResponseShouldMatchJson)
-}
